Change overhead rate, revamp instructions
Overhead rate was set back to last audited rate, as this should be used in place of the projected overhead rate for 3C and SPR work.
Also, the first-pass edited directions were substituted for the existing instruction text, and placed in text boxes instead of broken up awkwardly in regular Excel cells.